Added aria-hidden to SVG
Adding aria-hidden to SVGs to suppress Level Access errors. The images are decorative anyway, so this is fine. aria-hidden should not interfere with global aria-hidden set to display: none as SVG's are set to block, which overrides the global default. If nay issues, lmk, but do not think there will be a problem with this add. cc: @Brockenstein @rehamm01